feat: Add contacts to User model and seed data
This commit adds the "contacts" field to the User model, representing the list of users that the current user has added as contacts. It also adds the "addedContacts" field to represent the list of users who have added the current user as a contact. The schema.prisma file has been updated accordingly. Additionally, seed data for contacts has been added in the seed.js file.